requirements.txt
# --- Part 1: Weather Data Viewer (Flask) ---
Flask==3.0.3
requests==2.31.0
python-dotenv==1.0.1

# --- Part 2: Python CSV Plotter (Streamlit) ---
streamlit==1.32.0
pandas==2.2.1
plotly==5.19.0

# --- AI Integration Layer ---
openai==1.14.0
# Or google-generativeai==0.4.1 (depending on your chosen LLM)

# --- Utility & Portability ---
gunicorn==21.2.0
openpyxl==3.1.2